implementation of 'fcntl' locking backend
The idea behind introducing new backend was that default
'flock' backend has a disadvantage (or advantage) of survive
across forking. It's potentialy dangerous situation in some cases.
For instance, imagine that IPC::CL::Withstandby is used for building HA pair
of daemons, and daemon does forks after acquiring a lock. Now, if the
main process dies before any child, the standby daemon will fail to
promote itself to main one, because the lock is still held by all
children. The solution to this case would be to either explicitly close
lock file handler after forking, or use another locking mechanism which
is free of this feature.

However, fcntl behaivour is very nasty. So, you should be really
carefull in using it.